好崩溃，后端改好了除了模拟功能都测试好了，到底还要改多少

+++ backend/src/main/java/com/weread/repository/BookshelfRepository.java
@@ -15,20 +15,20 @@ import java.util.Optional;
 public interface BookshelfRepository extends JpaRepository<BookshelfEntity, Integer> {
 
     // 1. Check if a book exists on the user's bookshelf
-    Optional<BookshelfEntity> findByUserIdAndBookId(Long userId, Integer bookId);
+    Optional<BookshelfEntity> findByUserIdAndBookId(Integer userId, Integer bookId);
 
     // 2. Query all bookshelf items for a specific user
-    List<BookshelfEntity> findByUserId(Long userId);
+    List<BookshelfEntity> findByUserId(Integer userId);
 
     // 3. Query bookshelf items filtered by status (unread/reading/finished)
-    List<BookshelfEntity> findByUserIdAndStatus(Long userId, String status);
+    List<BookshelfEntity> findByUserIdAndStatus(Integer userId, String status);
 
     // 4. Custom query to update a book's status and last read time
     @Modifying
     @Query("UPDATE BookshelfEntity b SET b.status = :status, b.lastReadAt = :lastReadAt " +
             "WHERE b.userId = :userId AND b.bookId = :bookId")
     void updateBookStatus(
-            @Param("userId") Long userId,
+            @Param("userId") Integer userId,
             @Param("bookId") Integer bookId,
             @Param("status") String status,
             @Param("lastReadAt") LocalDateTime lastReadAt);
